refactor: replace string literals with CacheType enum in caching annotations and rename BinaryContentCreatedEvent to FileStoreEvent

+++ src/main/java/com/sprint/mission/discodeit/domain/repository/MessageRepository.java
@@ -43,5 +43,5 @@ public interface MessageRepository extends JpaRepository<Message, UUID> {
         SET m.author = null
         WHERE m.author.id = :userId
         """)
-    int nullifyAuthorByUserId(@Param("userId") UUID userId);
+    long nullifyAuthorByUserId(@Param("userId") UUID userId);
 }
